feat: hd08 所有已有策略都验证通过。d3_opt2 对 hd08 不存在（hd08 只有 1 个输出，只有 d1a/d1b/d2 三种策略）。 修复总结： 根因是  中  类型为 （32 位），但 hd07 d3_opt2 的合并 z 空间有 m=57 个变量。当解析 .poly 文件中位位置 ≥32 的单项式时， 行为未定义，导致高 25 位被丢弃，ANF 求值错误。 修改了三处： 1. :  →  2. :  →   →  3. :  →  4. main 验证循环:  →   →  现在所有输出穷举验证（256/256）通过。要继续什么？这个验证 bug 修复后，我们可以运行完整的优化管线。

+++ cpp/src/verify_anf.cpp
@@ -30,7 +30,7 @@ struct AffineTransform {
 };
 
 struct PolyTerm {
-    uint32_t mask;   // which z-variables appear (bits 0..m-1)
+    uint64_t mask;   // which z-variables appear (bits 0..m-1)
     int coeff;       // 0 or 1
 };
 
@@ -95,10 +95,10 @@ static std::vector<PolyOutput> parse_poly(const std::string& path, int& m) {
         for (int ti = 0; ti < term_counts[oi]; ti++) {
             // Parse [e_0, e_1, ..., e_{m-1}, coeff]
             char c; f >> c; // skip '['
-            uint32_t mask = 0;
+            uint64_t mask = 0;
             for (int b = 0; b < m; b++) {
                 int bit; f >> bit; f >> c; // read bit and comma/']
-                if (bit) mask |= (1u << b);
+                if (bit) mask |= (1ULL << b);
             }
             int coeff; f >> coeff;
             f >> c; // skip ']'
@@ -112,7 +112,7 @@ static std::vector<PolyOutput> parse_poly(const std::string& path, int& m) {
 //  Evaluate ANF at given z-value
 // ====================================================================
 
-static bool eval_poly(const PolyOutput& out, uint32_t z) {
+static bool eval_poly(const PolyOutput& out, uint64_t z) {
     bool result = false;
     for (const auto& term : out.terms) {
         if (term.coeff == 0) continue;
@@ -211,11 +211,11 @@ int main(int argc, char** argv) {
     for (int test = 0; test < n_total; test++) {
         uint32_t x = (n <= 16) ? (uint32_t)test
                                : (uint32_t)(((uint64_t)test * 0x9e3779b97f4a7c15ULL) & mask);
-        // z = Mx+b
-        uint32_t z = 0;
+        // z = Mx+b (z can be up to 64 bits for s > 32)
+        uint64_t z = 0;
         for (int j = 0; j < tfm.s; j++) {
             if ((__builtin_popcount(tfm.M_rows[j] & x) & 1) ^ tfm.b[j])
-                z |= (1u << j);
+                z |= (1ULL << j);
         }
 
         auto f_vals = eval_circuit_point(circ, x);
